content/canvas/test/test_2d.gradient.object.return.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.gradient.object.return</title>
<!-- Testing: createLinearGradient() and createRadialGradient() returns objects implementing CanvasGradient -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

window.CanvasGradient.prototype.thisImplementsCanvasGradient = true;

var g1 = ctx.createLinearGradient(0, 0, 100, 0);
ok(g1.addColorStop !== undefined, "g1.addColorStop !== undefined");
ok(g1.thisImplementsCanvasGradient === true, "g1.thisImplementsCanvasGradient === true");

var g2 = ctx.createRadialGradient(0, 0, 10, 0, 0, 20);
ok(g2.addColorStop !== undefined, "g2.addColorStop !== undefined");
ok(g2.thisImplementsCanvasGradient === true, "g2.thisImplementsCanvasGradient === true");

SimpleTest.finish();

});
</script>
